Changes to allow full customization of the Reaction application container structure (#2874).
