+++ imports/plugins/core/router/client/browserRouter.js
@@ -9,10 +9,10 @@ import { isEqual } from "lodash";
 import queryParse from "query-parse";
 import { Session } from "meteor/session";
 import { Tracker } from "meteor/tracker";
-import App from "/imports/plugins/core/router/client/app";
 import { Router } from "../lib";
 import { MetaData } from "/lib/api/router/metadata";
 import { TranslationProvider } from "/imports/plugins/core/ui/client/providers";
+import { Components } from "@reactioncommerce/reaction-components";
 
 const history = Router.history;
 
@@ -152,7 +152,7 @@ export function initBrowserRouter() {
       ReactDOM.render((
         <BrowserRouter history={history}>
           <TranslationProvider>
-            <App children={Router.reactComponents} />
+            <Components.App children={Router.reactComponents} />
           </TranslationProvider>
         </BrowserRouter>
       ), getRootNode());
