+++ imports/plugins/core/ui/client/containers/index.js
@@ -1,5 +1,6 @@
 export { default as EditContainer } from "./edit";
 export { default as Alerts } from "./alerts";
+export { default as App } from "./appContainer";
 export { default as ReactionAvatar } from "./avatar";
 export { default as SortableItem } from "./sortableItem";
 export { default as MediaGalleryContainer } from "./mediaGallery";
